+++ Samples~/Pokemon/Scripts/PokemonApi.cs
@@ -1,0 +1,16 @@
+﻿using KindMen.Uxios.Api;
+using KindMen.Uxios.PokemonObjects;
+using RSG;
+
+namespace KindMen.Uxios
+{
+    public class PokemonApi
+    {
+        public IPromise<Pokemon> GetPokemon(string name)
+        {
+            var pokemonResource = Resource<Pokemon>.At("https://pokeapi.co/api/v2/pokemon/" + name);
+
+            return pokemonResource.Value;
+        }
+    }
+}