+++ Samples~/Pokemon/Scripts/PokemonObjects/Pokemon.cs
@@ -1,0 +1,14 @@
+﻿namespace KindMen.Uxios.PokemonObjects
+{
+    public class Pokemon
+    {
+        public int id;
+        public string name;
+        public int base_experience;
+        public int height;
+        public bool is_default;
+        public int order;
+        public int weight;
+        public PokemonSprites sprites;
+    }
+}