+++ Samples~/Pokemon/Scripts/PokemonObjects/PokemonSprites.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace KindMen.Uxios.PokemonObjects
+{
+    public class PokemonSprites
+    {
+        public Uri front_default;
+    }
+}