+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Using the Sprite generic to translate URLs into Sprites work now
+
+### Added
+
+- Added a new Sample to demonstrate basic usage to tie Uxios onto a UI
+
 ## [0.1.0] - 2024-11-06
 
 ### Added